cri: spec_linux: factor out cgroup1/cgroup2 specifics
